refactor: move DateFormatter from presentation to domain layer
